// apps/api/src/modules/user/activity/application/activity.dto.ts
export type RecordUserActivityPayloadDTO = {
  userId: string;

  category: ActivityCategory;
  type: ActivityType;

  title: string;
  subtitle?: string;

  xpAwarded?: number;
  xpBucket?: ActivityXpBucket;
  coinsAwarded?: number;

  eventKey: string;

  trackerId?: string;
  topicId?: string;
  subtopicId?: string;
  mockTestId?: string;
  attemptId?: string;
  sourceUserId?: string;

  details?: ActivityDetails;
  occurredAt?: Date;
  utcOffsetMinutes?: number;
};